Add LAYOUT_HIDE_NAVIGATION to navigation bar
Quantum theme moves the bar if LAYOUT_HIDE_NAVIGATION is not
specified. Note that because of the way the framework is designed,
this have the side effect of also applying LAYOUT_FULLSCREEN
so contents may be behind the status bar. This will be fixed later
when we enable the transparent status bar.

Change-Id: I2403571711678f457da07b1e4645df1931612530