[SetupWizardLib] Implement SetupWizardLayout
SetupWizardLayout is a generic implementation of a "chrome" for the
activity. The default implementation uses suw_template, but the client
is free to use an alternative layout using the android:layout
attribute.

SetupWizardLayoutHelper is designed to work with suw_template and its
variants. Custom templates can implement their own LayoutHelpers to
perform layout specific tasks on the template.

Bug: 18685749
Change-Id: Ib94602b62cd8faba41f171114651a54a90c5ee8d