[SetupWizardLib] Add progress bar support
Added methods showProgressBar, hideProgressBar and isProgressBarShown
to SetupWizardLayout, that will show the bar if the corresponding
ViewStub exists.

This implementation has an improvement over existing Setup Wizard
implementations in that it supports ListViews as well as ScrollViews,
and the progress bar will be sticky together with the header instead
of being scrolled away.

Change-Id: Ifd7acae36a9b244d759b6528bb5dfdc3e1d15091