[SetupWizardLib] Fix Partner class and tests
Partner.java had a bug preventing it from fetching the default value
from the original package if the resource is not found in the
overlay. There is a corresponding issue in the test, which is using
its own resources for testing. Since the test package builds against
the source of the library, all resources in the library is also
available in the test package, and therefore the behavior of resource
not found in overlay package could not be tested. This CL fixes that
by using a MockResources object instead.

In addition, the attribute isOverlay was added to ResourceEntry so
clients can tell if the ResourceEntry came from the overlay or from
the fallback.

Change-Id: Ib0490277db502ea563171d79ab8491934dc43d01