Create a standalone gradle file without SDK/build-tools.
Currently, setup-wizard-lib has two gradle files - build.gradle,
which is used when building from inside the Android tree, where
all dependencies are built from source, and standalone.gradle,
where all dependencies are prebuilts. Unbundled apps generally
want the latter, but this gradle file defines a specific build
tools and SDK version, which is undesirable if we want to build
everything with the same build tools and SDK.

So, introduce a third gradle file, standalone-rules.gradle, which
uses prebuilt support lib dependencies, but which does not define
the SDK and build tools so that another gradle file can define
these at the top level.

Since this doesn't affect existing gradle files, this should be
a safe change.

Bug: 22670848
Change-Id: I95f4294105b54c5d978f6632bd98e2492a6d983c