[SuwLib] Implement Items framework
Items framework is a framework modeled after preferences, which uses
XML to inflate a collection of items. Instead of using activity or
fragment directly like preferences, a ListAdapter is created and you
can use that with any existing ListViews.

SetupWizardItemsLayout is a convenient wrapper around
SetupWizardListLayout which will automatically inflates
android:entries in its attributes to populate the list.

Note: A recycler view adapter is under consideration

Change-Id: I5eb8853c160cf86fa5b6f21a01dfa4b0030643f6