[SetupWizard] Adds a new divider style for showing divider.
This style is used to determine if the divider item decoration should
check both the item view holder below and above where the divider would
draw itself. If set to true, both need to be true. By default, this is
set to either, so that a divider is drawn if either the item below or
above the divider allows a divider to be set.

bug:27619572
Change-Id: Ic0494542b3b62bf19061fb9feda52baac90f3972