Refactor GenericInflater
Split GenericInflater into SimpleInflater, which wraps the XML parsing
into finding the tag name and an AttributeSet. The subclass is
responsible for creating the corresponding object for the given tag
name and AttributeSet.
ReflectionInflater is a subclass of SimpleInflater which will treat
the tag name as a class, and create the object via reflection using
the 2-argument constructor TagName(Context, AttributeSet).

Test: ./gradlew connectedAndroidTest
Change-Id: I81408f099f132b7bedb27a9b5425e3c18988aa52