Add workaround for touch event propagation
In TextView with links, onTouchEvent always return true regardless of
return value from the movement method because View#onTouchEvent is
hard coded to return true whenever the view is clickable.

In RichTextView, add a layer of abstraction to allow the movement
method to override the return value. TouchableLinkMovementMethod is
an implementation added, which will consume the event only when a
link is being touched.

Test: ./gradlew test
Bug: 77338508
Change-Id: I761579d6f153a41beb979acd88a4090eb3998cf3